+++ src/exchanges/ftx/streams/accountDataStream.js
@@ -65,30 +65,6 @@ class AccountDataStream extends FtxWebsocketBase {
 
         this.emit('orderUpdate', order);
     }
-
-    transformFtxOrder(fOrder) {
-        /**
-         * @type {PlacedOrder}
-         */
-        const zOrder = {};
-
-        zOrder.id = fOrder.id.toString();
-        zOrder.timestamp = Date.parse(fOrder.createdAt);
-        zOrder.symbol = fOrder.market;
-        zOrder.type = fOrder.type;
-        zOrder.side = fOrder.side;
-        zOrder.quantity = parseFloat(fOrder.size);
-        zOrder.price = fOrder.price ? parseFloat(fOrder.price) : undefined;
-        // zOrder.trades = bOrder.fills; // TODO: Fill commision counter
-
-        if (fOrder.status === 'new') {
-            zOrder.status = 'open';
-        } else {
-            zOrder.status = fOrder.status;
-        }
-
-        return zOrder;
-    }
 }
 
 module.exports = AccountDataStream;
